feat: esign repo importing support
Esign repo importing support - this required extracting the esign decryption key from kravasign and reverse engineering how they decrypt the codes to be able to see what repos are listed.

ESign has no reason to be doing ANY of this apart from trying to be a pain in the ass for its users, however we got it done after a long time.

Co-Authored-By: HAHALOSAH <67280050+HAHALOSAH@users.noreply.github.com>
Co-Authored-By: nekohaxx <151578517+nekohaxx@users.noreply.github.com>