Ability to running in the background
This code change is used to play a 10 seconds silent audio file when Feather is:
	•	Performing the “Import from URL” process
	•	Downloading any IPAs from any Repos
	•	Importing
	•	Packaging
	•	Sending Payload

Feather will continue to run in the background during these processes. Even when the device is locked, Feather will keep running & the audio will automatically stop once the processes are completed. I’ve tested it using an MP3 file with actual sound and a fade-out at the end to confirm the audio loop behavior.