fix: don't rely on xcode compression
It's better to specify COMPRESS_PNG_FILES=YES in compiler flags. This is because Xcode compresses PNG images rather horribly; it tries to compress them in a way to make them better for rendering instead of size, but it's actually worse for rendering in practice since the extra time to read data on the NAND ends up being much more costly than what Xcode saves (which is just a few bitflips in the end...)

In this commit, I have specified Xcode to not compress PNG images, and have ran them through a lossless compressor so they maintain the same quality while still being better for storage.